initial commit: project setup & build code written
This gist of this project is that my dad has two formats of excel
documents and needs a program that can update a larger 'project'
spreadsheet based upon the revisions specified by a computer-generated
(by revit) spreadsheet with dates and ids.

A full spec may be written at a later date.

I've also included a simple READEME.md and also a LICENSE.

I've initialized a new zig project for this project, and I've also added
xlsxio as a dependency for this project.

For the xlsxio library, it also depended on expat and also minizip,
which also depended zlib. All of which, I have manually resolved and
added to the build.zig file so that they all compile and link
successfully.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>