added date parsing for revit spreadsheets
I've added the ability to parse dates, this is in anticipation of
parsing the specified dates provided within the revit-generated
spreadsheets.
The dates themselves are stored as arrays in the format of
`yyyy/mm/dd` to be country agnostic.
Tests have also been written.

I've also removed some old testing code from the build.zig file that
is no longer required.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>