added dependency libxls & updated .gitignore
I've added the C library libxls as a dependency as the actual document
I'm supposed to be modifying is of the format .xls which is different
from the .xlsx format and cannot be opened by xlsxio due to one using
XML as a backend and the other use OLE.

There is a slight, and by slight, I mean major issue where there are no
FOSS libraries capable of modifying a .xls document which is the entire
point of this program. So, as a comprimise, this program will instead
generate a new document with the necessary changes in the format of
.xlsx and then let the user copy the changes without formatting and
paste into the original document, so as to delegate the work of
modification to the official excel program.

I've also updated the .gitignore to include some other files that aren't
meant to be included in the git repository as they are files for
testing.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>