revit parsing: added basic parsing and listing
I've written a basic parser that's capable of parsing the entire revit
file into separate rows. And I've written some basic code to test the
parsing of those revit generated files and list the rows parsed.

This works fine for now, though I will change some basic aspects of how
it works, for instance, the id won't be included with the row data
anymore, and instead of returning a row, the parseRow function will
simply insert into a hashmap at that speicifed id with the rest of the
row data.

The format of the row is <id> <name> <?revision> <?date>.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>