revit rows: moved to appending to an arraylist
I've chosen to use an arraylist to hold the parsed revit rows as there
won't be many revit rows so an arraylist will be much faster due to
cpu caching. And I've chosen to simply insert into the arraylist when
I parse as there will be no instance in which I won't store it in the
array.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>