monolith: added parsing of monolith dates
I've written some basic parsing of the XLS monolith dates, as an array
of allocated dates for reference when updated the revision values.

Also the date has been moved to the main.zig file as it is no longer
exclusive to revit.zig.

I've also added the .~lock files to the gitignore so any office backups
aren't included in this repo.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>