conversion from monolith to output written
I've chosen to do the modifications through simply copying the contents
of the monolith to an output xlsx file and checking upon every new row
if the 'id' matches one defined by the revit file. If so, then copy
all of the contents up until that coloumn, and then write the
modification (as in the revision) and then copy all of the contents
after the revision.

Modifications have also been made so that the dates used are normalised
by adding 2000 so that they are the same.

Test code has been written, and WORKS.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>